refactor(Number): tryCount int -> Number 객체로 변경
